Improve device detection for libinput 1.27
I have noticed on my own laptop that libinput-gestures sometimes picks
the wrong device. Since libinput 1.27, it seems that my USB receiver is
now also marked as a "gesture" device so it if is first in the device list
then it gets erroneously selected. Now we filter the list of devices
with gesture capability and prefer one which has touchpad or trackpad in
it's name.

Have tried to keep this change maximally backwards compatible to ensure
existing configurations are not broken.